+++ docs/changelog.rst
@@ -3,11 +3,11 @@
 Changelog
 =========
 
-Version 0.2.0 - Unreleased
+Version 0.2.0 - 2018/04/08
 --------------------------
 
 * Added Sass function ``floor-number-items()``;
-* Removed ``flat`` property in favor of ``structure`` to be able to allow other structure modes;
+* Removed ``flat`` property in favor of ``structure`` to allow other structure modes;
 * Added new structure mode ``list``;
 * Added new structure mode ``string``;